LPD-88830 Move the Key Manager modules to modules/apps/portal-security
