+++ build.properties
@@ -187,7 +187,6 @@
     #    apps/item-selector,\
     #    apps/journal,\
     #    apps/json-storage,\
-    #    apps/keymanager,\
     #    apps/knowledge-base,\
     #    apps/launch,\
     #    apps/layout,\
