--- modules/apps/keymanager/keymanager-api/bnd.bnd
@@ -1,5 +1,0 @@
-Bundle-Name: Liferay Key Manager API
-Bundle-SymbolicName: com.liferay.keymanager.api
-Bundle-Version: 1.0.0
-Export-Package: com.liferay.keymanager
-Liferay-Releng-Module-Group-Title: Key Manager
