--- modules/apps/keymanager/keymanager-impl/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Key Manager Implementation
-Bundle-SymbolicName: com.liferay.keymanager.impl
-Bundle-Version: 1.0.0
-Liferay-Releng-Module-Group-Title: Key Manager
