--- modules/apps/keymanager/keymanager-impl/src/main/resources/META-INF/module-log4.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0"?>
-
-<Configuration strict="true">
-	<Loggers>
-		<Logger level="INFO" name="com.liferay.keymanager.internal" />
-	</Loggers>
-</Configuration>
