--- modules/apps/keymanager/keymanager-spi/build.gradle
@@ -1,4 +1,0 @@
-dependencies {
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly project(":apps:keymanager:keymanager-api")
-}
