+++ modules/apps/portal-security/portal-security-key-impl/src/main/resources/META-INF/module-log4.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0"?>
+
+<Configuration strict="true">
+	<Loggers>
+		<Logger level="INFO" name="com.liferay.portal.security.key.internal" />
+	</Loggers>
+</Configuration>
