+++ modules/apps/portal-security/portal-security-key-spi/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
+	compileOnly project(":apps:portal-security:portal-security-key-api")
+}
