+++ modules/apps/portal-security/portal-security-key-spi/src/main/resources/com/liferay/portal/security/key/spi/profile/packageinfo
